requirements.txt
Django==2.2.10
djangorestframework==3.11.0
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
uritemplate==3.0.1
